requirements.txt
# Core Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
# torch-scatter>=2.1.0  # Optional - commented out due to Windows build issues
# torch-sparse>=0.6.0   # Optional - commented out due to Windows build issues  
# torch-cluster>=1.6.0  # Optional - commented out due to Windows build issues
# Note: PyG will use fallback implementations without these (slightly slower but functional)

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
pyverilog>=1.3.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
loguru>=0.7.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
